Refactor `Phoenix` binding/config logic.
