+++ src/config.ts
@@ -1,4 +1,4 @@
-import ConfigSchema, { defaultConfig } from './utils/ConfigSchema';
+import ConfigSchema, { defaultConfig } from './model/ConfigSchema';
 
 const CONFIG: ConfigSchema = Object.assign({}, defaultConfig, {
   DEBUG_ENABLED: true,
