+++ src/model/ConfigSchema.ts
@@ -1,6 +1,6 @@
-import ContainerNode from '../nodes/ContainerNode';
-import TileNode from '../nodes/TileNode';
-import SyncedFrame from '../frames/SyncedFrame';
+import ContainerNode from './nodes/ContainerNode';
+import TileNode from './nodes/TileNode';
+import SyncedFrame from './frames/SyncedFrame';
 
 export default interface ConfigSchema {
   DEFAULT_CONTAINER: new (syncedFrame: SyncedFrame) => ContainerNode;
