+++ src/phoenix/config.ts
@@ -1,0 +1,7 @@
+export default interface ConfigSchema {
+  DEBUG_JSON_SPACE: string | number;
+}
+
+export default <ConfigSchema>{
+  DEBUG_JSON_SPACE: 2,
+};
