+++ src/phoenix/index.ts
@@ -1,3 +1,2 @@
 import initialise from './initialise';
-
 initialise();
